This avoids frequently opening and parsing and merging configurations that do not change for long-long time.

The change is trying to minimize breakages (your config will still be reloaded in 60 seconds) while avoiding the penalty of loading config on every request as much as possible.


Some test breakage is expected because of how they are written, but nothing real should break for long time. 


| Benchmark                      | before  | after                 |
|--------------------------------|:-------:|:---------------------:|
| web_requests_request_overheads | 2.58 ms | 2.45 ms: 1.05x faster |
| load_getdoc_synthetic          | 183 rps | 189 rps: 1.02x faster |


TODO:

- [x] ways to quickly reload? `SIGHUP` ?
- [x] Bounds on site_cache size? Right now it doesn't seem like the maxsize is global. fixed: https://github.com/frappe/frappe/pull/28939